% Generated by roxygen2 (4.0.2): do not edit by hand
\name{backtransform}
\alias{backtransform}
\title{Back-transformations}
\usage{
backtransform(x, type = c("identity", "log", "logit", "none", NA_character_))
}
\arguments{
\item{x}{value to back-transform}

\item{type}{type of transform (log, logit).}
}
\value{
backtransformed value
}
\description{
Performs inverse log or logit transformations.
}

