% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{EQR}
\alias{EQR}
\title{Annual Average Environmental Quality Ratio for Macrofauna.}
\usage{
data(EQR)
}
\description{
Annual artithmetic average environmental quality ratio's (EQR) for Macrofauna
in a specific water body, based on Baggelaar et al., (2010)
}
\details{
The columns represent the following information:
 \itemize{
     \item{OBJECTID: water body code}
     \item{PAR: parameter, in this case EQR;}
     \item{color: colors in density function;}
     \item{DATE: year;}
     \item{VALUE: annual arithmetic average EQR}
     \item{TARGET: target EQR;}
     \item{transfrom: applied transform}
 }
}
\references{
Baggelaar, P., O. van Tongeren, R. Knoben, & W. van Loon, 2010.
     Rapporteren van de betrouwbaarheid van KRW-beoordelingen. H2O 16, p.21--25
}
\keyword{datasets}

