% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mya}
\alias{mya}
\title{Multi-Year Average}
\usage{
mya(x, ...)
}
\arguments{
\item{x}{object of class \code{\link{conf_input}} or a \code{\link{data.frame}}
that can be coerced to an instance of class \code{\link{conf_input}}.}

\item{\dots}{further arguments to be passed to other methods}
}
\value{
a \code{\link{data.frame}} with the following columns:
\itemize{
    	\item{\code{MYA}: }{the multi-year arithmetic average;}
 	\item{\code{PROB_LTT}: }{the probability that \code{MYA} is less than the target value specified;}
 	\item{\code{PROB_GTT}: }{the probability that \code{MYA} is greater than the target value specified;}
     \item{\code{q05}: }{the lowerbound of the 90\% confidence interval of \code{MYA}}
     \item{\code{q95}: }{the upperbound of the 90\% confidence interval of \code{MYA}}
 }
}
\description{
Estimates the multi-year average of environmental properties and
associated confidence intervals.
}
\seealso{
\code{\link{conf}}
}
