% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strip_spaces}
\alias{strip_spaces}
\title{Remove Redundant Spaces}
\usage{
strip_spaces(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without trailing or multiple spaces
}
\description{
This function removes redundant spaces from character vectors
}
\examples{
stopifnot(confidence:::strip_spaces(" Hello  World  ") == "Hello World")
}

