% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform}
\alias{transform}
\title{Transformations
 
Performs log or logit transformations.}
\usage{
transform(x, type = c("identity", "log", "logit", "none", NA_character_))
}
\arguments{
\item{x}{value to transform}

\item{type}{type of transform (log, logit).}
}
\value{
transformed value
}
\description{
Transformations
 
Performs log or logit transformations.
}
