% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_html}
\alias{write_html}
\title{Internal Functions For Writing HTML}
\usage{
write_html(x, ...)
}
\arguments{
\item{x}{object}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Internal Functions For Writing HTML
}
