% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DCA}
\alias{DCA}
\title{Annual Average 1,2-dichloroethane Concentration}
\source{
Baggelaar, P., O. van Tongeren, R. Knoben, & W. van Loon, 2010. 
   Rapporteren van de betrouwbaarheid van KRW-beoordelingen. H2O 16, p.21--25
}
\description{
Annual arithmetic average concentration of 1,2-dichloroethane (DCA) in
a specific water body (\eqn{\mu g/l}{ug/l}), based on Baggelaar et al., (2010)
}
\details{
The columns represent the following information:
\describe{
  \item{OBJECTID}{water body code}
  \item{PAR}{parameter, in this case 1,2-dichloroethane}
  \item{color}{colors in density function}
  \item{DATE}{year}
  \item{VALUE}{annual arithmetic average concentration}
  \item{TARGET}{target according to the European Water Framework Directive}
  \item{UNIT}{measurement unit (\eqn{\mu g/l}{ug/l})}
}
}
