% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize}
\alias{sanitize}
\title{Sanitize Text to Give Proper Filenames}
\usage{
sanitize(x)
}
\arguments{
\item{x}{character vector to sanitize}
}
\value{
sanitized character vector
}
\description{
Sanitize Text to Give Proper Filenames
}
