% Generated by roxygen2 (4.0.2): do not edit by hand
\name{transform}
\alias{transform}
\title{Transformations}
\usage{
transform(x, type = c("identity", "log", "logit", "none", NA_character_))
}
\arguments{
\item{x}{value to transform}

\item{type}{type of transform (log, logit).}
}
\value{
transformed value
}
\description{
Performs log or logit transformations.
}

